change structure
tgbot with handlers are in separate folder. separate commands, dispatcher in tgbot
