+++ tgbot/handlers/broadcast_message/handlers.py
@@ -9,8 +9,8 @@ from .manage_data import CONFIRM_DECLINE_BROADCAST, CONFIRM_BROADCAST
 from .keyboards import keyboard_confirm_decline_broadcasting
 from .static_text import broadcast_command, broadcast_wrong_format, broadcast_no_access, error_with_html, \
     message_is_sent, declined_message_broadcasting
-from tgbot.models import User
-from tgbot.tasks import broadcast_message
+from users.models import User
+from users.tasks import broadcast_message
 
 
 def broadcast_command_with_message(update: Update, context: CallbackContext):
